requirements.txt
Flask==3.1.1
Jinja2==3.1.6
python-decouple==3.8
